requirements.txt
psycopg2-binary
python-dotenv
ollama
pandas
pgvector
playwright
playwright-stealth
